requirements.txt
asgiref==3.2.7
astroid==2.4.2
autopep8==1.5.4
boto3==1.14.56
botocore==1.17.56
certifi==2020.4.5.2
chardet==3.0.4
coverage==5.1
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-filter==2.2.0
django-rest-auth==0.9.5
django-storages==1.10
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
djangorestframework-stubs==1.2.0
docutils==0.15.2
idna==2.9
inflection==0.4.0
isort==5.6.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyline==0.3.20
pylint==2.6.0
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
